refactor useEntryAddition
